+++ src/components/form/EditModeField/stories/EditModeField.stories.tsx
@@ -57,9 +57,6 @@ export default {
     disableCopying: {
       control: { type: 'boolean' },
     },
-    displayClearIcon: {
-      control: { type: 'boolean' },
-    },
     displayInline: {
       control: { type: 'boolean' },
     },
@@ -70,9 +67,6 @@ export default {
       control: { type: 'boolean' },
     },
     skeleton: {
-      control: { type: 'boolean' },
-    },
-    isLoading: {
       control: { type: 'boolean' },
     },
     multilineView: {